requirements.txt
langchain==0.3.25
langchain_community==0.3.25
langchain_core==0.3.65
langchain_google_genai==2.1.5
pydantic==2.11.7
python-dotenv==1.1.0
python_docx==1.2.0
qdrant_client==1.14.3
streamlit==1.46.0
